+++ boards/ark/fmu-v6xrt/src/CMakeLists.txt
@@ -60,7 +60,6 @@ else()
 	endif()
 
 	px4_add_library(drivers_board
-		autoleds.c
 		automount.c
 		#can.c
 		i2c.cpp
