+++ boards/ark/fmu-v6xrt/src/led.c
@@ -34,7 +34,7 @@
 /**
  * @file led.c
  *
- * PX4 fmu-v6rt LED backend.
+ * ARK V6X-RT LED backend.
  */
 
 #include <px4_platform_common/px4_config.h>
@@ -45,6 +45,7 @@
 #include <hardware/imxrt_gpio.h>
 #include "board_config.h"
 
+#include <nuttx/board.h>
 #include <arch/board/board.h>
 
 /*
@@ -61,13 +62,30 @@ extern void led_off(int led);
 extern void led_toggle(int led);
 __END_DECLS
 
-
+#ifdef CONFIG_ARCH_LEDS
+static bool nuttx_owns_leds = true;
+//                                B  R  S  G
+//                                0  1  2  3
+static const uint8_t xlatpx4[] = {1, 2, 4, 0};
+#  define xlat(p) xlatpx4[(p)]
 static uint32_t g_ledmap[] = {
-	GPIO_nLED_BLUE,   // Indexed by LED_BLUE
-	GPIO_nLED_RED,    // Indexed by LED_RED, LED_AMBER
-	GPIO_LED_SAFETY,  // Indexed by LED_SAFETY
-	GPIO_nLED_GREEN,  // Indexed by LED_GREEN
+	GPIO_nLED_GREEN,   // Indexed by BOARD_LED_GREEN
+	GPIO_nLED_BLUE,    // Indexed by BOARD_LED_BLUE
+	GPIO_nLED_RED,     // Indexed by BOARD_LED_RED
+	GPIO_nSAFETY_SWITCH_LED_OUT,  // Indexed by LED_SAFETY by xlatpx4
 };
+
+#else
+
+#  define xlat(p) (p)
+static uint32_t g_ledmap[] = {
+	GPIO_nLED_BLUE,                     // Indexed by LED_BLUE
+	GPIO_nLED_RED,                      // Indexed by LED_RED, LED_AMBER
+	GPIO_LED_SAFETY,                    // Indexed by LED_SAFETY
+	GPIO_nLED_GREEN,                    // Indexed by LED_GREEN
+};
+
+#endif
 
 __EXPORT void led_init(void)
 {
@@ -90,7 +108,7 @@ static void phy_set_led(int led, bool state)
 
 static bool phy_get_led(int led)
 {
-
+	/* If Low it is on */
 	if (g_ledmap[led] != 0) {
 		return !imxrt_gpio_read(g_ledmap[led]);
 	}
@@ -100,16 +118,119 @@ static bool phy_get_led(int led)
 
 __EXPORT void led_on(int led)
 {
-	phy_set_led(led, true);
+	phy_set_led(xlat(led), true);
 }
 
 __EXPORT void led_off(int led)
 {
-	phy_set_led(led, false);
+	phy_set_led(xlat(led), false);
 }
 
 __EXPORT void led_toggle(int led)
 {
-
-	phy_set_led(led, !phy_get_led(led));
+	phy_set_led(xlat(led), !phy_get_led(xlat(led)));
 }
+
+#ifdef CONFIG_ARCH_LEDS
+/****************************************************************************
+ * Public Functions
+ ****************************************************************************/
+
+/****************************************************************************
+ * Name: board_autoled_initialize
+ ****************************************************************************/
+
+void board_autoled_initialize(void)
+{
+	led_init();
+}
+
+/****************************************************************************
+ * Name: board_autoled_on
+ ****************************************************************************/
+
+void board_autoled_on(int led)
+{
+	if (!nuttx_owns_leds) {
+		return;
+	}
+
+	switch (led) {
+	default:
+		break;
+
+	case LED_HEAPALLOCATE:
+		phy_set_led(BOARD_LED_BLUE, true);
+		break;
+
+	case LED_IRQSENABLED:
+		phy_set_led(BOARD_LED_BLUE, false);
+		phy_set_led(BOARD_LED_GREEN, true);
+		break;
+
+	case LED_STACKCREATED:
+		phy_set_led(BOARD_LED_GREEN, true);
+		phy_set_led(BOARD_LED_BLUE, true);
+		break;
+
+	case LED_INIRQ:
+		phy_set_led(BOARD_LED_BLUE, true);
+		break;
+
+	case LED_SIGNAL:
+		phy_set_led(BOARD_LED_GREEN, true);
+		break;
+
+	case LED_ASSERTION:
+		phy_set_led(BOARD_LED_RED, true);
+		phy_set_led(BOARD_LED_BLUE, true);
+		break;
+
+	case LED_PANIC:
+		phy_set_led(BOARD_LED_RED, true);
+		break;
+
+	case LED_IDLE : /* IDLE */
+		phy_set_led(BOARD_LED_RED, true);
+		break;
+	}
+}
+
+/****************************************************************************
+ * Name: board_autoled_off
+ ****************************************************************************/
+
+void board_autoled_off(int led)
+{
+	if (!nuttx_owns_leds) {
+		return;
+	}
+
+	switch (led) {
+	default:
+		break;
+
+	case LED_SIGNAL:
+		phy_set_led(BOARD_LED_GREEN, false);
+		break;
+
+	case LED_INIRQ:
+		phy_set_led(BOARD_LED_BLUE, false);
+		break;
+
+	case LED_ASSERTION:
+		phy_set_led(BOARD_LED_RED, false);
+		phy_set_led(BOARD_LED_BLUE, false);
+		break;
+
+	case LED_PANIC:
+		phy_set_led(BOARD_LED_RED, false);
+		break;
+
+	case LED_IDLE : /* IDLE */
+		phy_set_led(BOARD_LED_RED, false);
+		break;
+	}
+}
+
+#endif /* CONFIG_ARCH_LEDS */
